Made changes to have token generator in same file
