+++ playwright/api-requests/Token.tsx
@@ -56,7 +56,7 @@ export class Token {
    * Used for fetching an Altinn test token for a specific role
    * @returns The Altinn test token as a string
    */
-  public async getPersonalAltinnToken(): Promise<string> {
+  public async getPersonalAltinnToken(partyUuid: string = 'asdads'): Promise<string> {
     const url =
       `https://altinn-testtools-token-generator.azurewebsites.net/api/GetPersonalToken?env=${process.env.environment}` +
       `&pid=${process.env.PID}` +
@@ -85,5 +85,34 @@ export class Token {
       throw new Error(`Failed to fetch token: ${response.statusText} - ${errorMessage}`);
     }
     return response.text();
+  }
+
+  public async generateAltinnPersonalToken(): Promise<string> {
+    const url =
+      `https://altinn-testtools-token-generator.azurewebsites.net/api/GetPersonalToken` +
+      `?env=${process.env.ENV_NAME}` +
+      `&pid=${process.env.DAGL_PID}` +
+      `&userid=${process.env.DAGL_USER_ID}` +
+      `&partyid=${process.env.DAGL_PARTY_ID}` +
+      `&partyuuid=${process.env.DAGL_PARTY_UUID}` +
+      `&authLvl=3&ttl=3000&scopes=altinn:portal/enduser`;
+
+    const authHeader = Buffer.from(`${this.username}:${this.password}`).toString('base64');
+
+    const response = await fetch(url, {
+      headers: { Authorization: `Basic ${authHeader}` },
+    });
+
+    const token = await response.text();
+
+    if (!token || !token.startsWith('ey')) {
+      throw new Error('Invalid token received from Altinn');
+    }
+
+    return token;
+  }
+  catch(err: unknown) {
+    console.error('Error retrieving Altinn token:', err);
+    throw err;
   }
 }
